Move health related things into sub-package
